requirements.txt
tensorboard

torch
numpy
Pillow
torchvision
segments-ai